==> 01_data-process.ipynb <==
<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Setup

<div style="background-color:teal; color:white; padding:5px; font-size:20px">
Installs

!pip install hopsworks

<div style="background-color:teal; color:white; padding:5px; font-size:20px">
Imports

import pandas as pd
import numpy as np

import importlib
import sys
sys.path.append('./scripts')

import feat_engineer_helpers

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Hopsworks Feature Store Connection

import boto3

# Initialize a client for Systems Manager
ssm = boto3.client('ssm', region_name='us-east-1')

parameter_name = 'hopsworks-api-key'

# Fetch the parameter
response = ssm.get_parameter(Name=parameter_name, WithDecryption=True)

# Extract the parameter value (API key in this case)
hopsworks_api_key = response['Parameter']['Value']

import hopsworks

project = hopsworks.login()

fs = project.get_feature_store()

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Data Download & Feature Engineering

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Articles

articles_df = pd.read_parquet('https://repo.hops.works/dev/jdowling/articles.parquet')
print(articles_df.shape)
articles_df.head(3)

# Check for NaNs
articles_df.isna().sum()[articles_df.isna().sum() > 0]

from feat_engineer_helpers import prepare_articles

articles_df = prepare_articles(articles_df)
articles_df.head(3)

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Customers

customers_df = pd.read_parquet('https://repo.hops.works/dev/jdowling/customers.parquet')
print(customers_df.shape)
customers_df.head(3)

# Check for NaNs
customers_df.isna().sum()[customers_df.isna().sum() > 0]

from feat_engineer_helpers import prepare_customers

customers_df = prepare_customers(customers_df)
print(len(customers_df))
customers_df.head(3)

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Transactions

trans_df = pd.read_parquet('https://repo.hops.works/dev/jdowling/transactions_train.parquet')[:600000]
print(trans_df.shape)
trans_df.head(3)

# Check for NaNs
trans_df.isna().sum()[trans_df.isna().sum() > 0]

from feat_engineer_helpers import prepare_transactions

trans_df = prepare_transactions(trans_df)
print(f"There are {len(trans_df):,} transactions in total.")
trans_df.head(3)

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Create Feature Groups (Hopsworks)

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Items

# Define Feature Group
articles_fg = fs.get_or_create_feature_group(
    name="articles",
    description="Fashion items data including type of item, visual description and category",
    version=1,
    primary_key=["article_id"],
    online_enabled=True
)
# Load Data
articles_fg.insert(articles_df)

# Add Feature Descriptions
feature_descriptions = [
    {"name": "article_id", "description": "Identifier for the article."},
    {"name": "product_code", "description": "Code associated with the product."},
    {"name": "prod_name", "description": "Name of the product."},
    {"name": "product_type_no", "description": "Number associated with the product type."},
    {"name": "product_type_name", "description": "Name of the product type."},
    {"name": "product_group_name", "description": "Name of the product group."},
    {"name": "graphical_appearance_no", "description": "Number associated with graphical appearance."},
    {"name": "graphical_appearance_name", "description": "Name of the graphical appearance."},
    {"name": "colour_group_code", "description": "Code associated with the colour group."},
    {"name": "colour_group_name", "description": "Name of the colour group."},
    {"name": "perceived_colour_value_id", "description": "ID associated with perceived colour value."},
    {"name": "perceived_colour_value_name", "description": "Name of the perceived colour value."},
    {"name": "perceived_colour_master_id", "description": "ID associated with perceived colour master."},
    {"name": "perceived_colour_master_name", "description": "Name of the perceived colour master."},
    {"name": "department_no", "description": "Number associated with the department."},
    {"name": "department_name", "description": "Name of the department."},
    {"name": "index_code", "description": "Code associated with the index."},
    {"name": "index_name", "description": "Name of the index."},
    {"name": "index_group_no", "description": "Number associated with the index group."},
    {"name": "index_group_name", "description": "Name of the index group."},
    {"name": "section_no", "description": "Number associated with the section."},
    {"name": "section_name", "description": "Name of the section."},
    {"name": "garment_group_no", "description": "Number associated with the garment group."},
    {"name": "garment_group_name", "description": "Name of the garment group."},
    {"name": "prod_name_length", "description": "Length of the product name."},
]

for desc in feature_descriptions: 
    articles_fg.update_feature_description(desc["name"], desc["description"])

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Customers

# Define Feature Group
customers_fg = fs.get_or_create_feature_group(
    name="customers",
    description="Customers data including age and postal code",
    version=1,
    primary_key=["customer_id"],
    online_enabled=True,  # Low latency access to data
)
# Load Data
customers_fg.insert(customers_df)

# Add Feature Descriptions
feature_descriptions = [
    {"name": "customer_id", "description": "Unique identifier for each customer."},
    {"name": "club_member_status", "description": "Membership status of the customer in the club."},
    {"name": "age", "description": "Age of the customer."},
    {"name": "postal_code", "description": "Postal code associated with the customer's address."},
    {"name": "age_group", "description": "Categorized age group of the customer."},
]

for desc in feature_descriptions:
    customers_fg.update_feature_description(desc["name"], desc["description"])

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Transactions

trans_fg = fs.get_or_create_feature_group(
    name="transactions",
    version=1,
    description="Transactions data including customer, item, price, sales channel and transaction date",
    primary_key=["customer_id", "article_id"],
    online_enabled=True,
    event_time="t_dat",
)
trans_fg.insert(
    trans_df,
    write_options={"wait_for_job": True},
)

feature_descriptions = [
    {"name": "t_dat", "description": "Timestamp of the data record."},
    {"name": "customer_id", "description": "Unique identifier for each customer."},
    {"name": "article_id", "description": "Identifier for the purchased article."},
    {"name": "price", "description": "Price of the purchased article."},
    {"name": "sales_channel_id", "description": "Identifier for the sales channel."},
    {"name": "year", "description": "Year of the transaction."},
    {"name": "month", "description": "Month of the transaction."},
    {"name": "day", "description": "Day of the transaction."},
    {"name": "day_of_week", "description": "Day of the week of the transaction."},
    {"name": "month_sin", "description": "Sine of the month used for seasonal patterns."},
    {"name": "month_cos", "description": "Cosine of the month used for seasonal patterns."},
]

for desc in feature_descriptions:
    trans_fg.update_feature_description(desc["name"], desc["description"])

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Create Rankings

# Load Feature Groups From Hopsworks
trans_fg = fs.get_feature_group(name="transactions",version=1,)
customers_fg = fs.get_feature_group(name="customers",version=1,)
articles_fg = fs.get_feature_group(name="articles",version=1,)

def compute_ranking_dataset(trans_fg, articles_fg, customers_fg):

    # DATASET JOINS & FILTERING
    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - 
    # Joined-Dataset Creation & Relevant Feature Filtering
    
    # Define the features used in the query
    query_features = ["customer_id", "age", "month_sin", "month_cos", "article_id"]
    
    # Perform the necessary joins to create the feature set
    fg_query = trans_fg.select(["month_sin", "month_cos"]).join(articles_fg.select_all(), on=["article_id"]).join(customers_fg.select(["customer_id", "age"]))
    df = fg_query.read()
    df = df[query_features]

    # NEG SAMPLE GENERATION
    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
    # Creates negative samples by randomly sampling from 3 sets:
    # article_id, customer_id & age-monthsin-monthcos feature chunks

    # Copy the positive pairs for ranking
    positive_pairs = df.copy()
    
    # Define the number of negative pairs to generate
    n_neg = len(positive_pairs) * 10

    # Create new id_list by sampling from unique list of positive samples
    negative_pairs = positive_pairs["article_id"].drop_duplicates().sample(n_neg, replace=True, random_state=2).to_frame()

    # Add customer_id to negative_pairs
    negative_pairs["customer_id"] = positive_pairs["customer_id"].sample(n_neg, replace=True, random_state=3).to_numpy()

    # Add other features to negative_pairs
    negative_pairs[["age", "month_sin", "month_cos"]] = positive_pairs[["age", "month_sin", "month_cos"]].sample(n_neg, replace=True, random_state=4).to_numpy()
    
    # Add labels to positive and negative pairs
    positive_pairs["label"] = 1
    negative_pairs["label"] = 0
    
    # FINISHING TOUCHES
    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
    
    # Concatenate positive and negative pairs
    ranking_df = pd.concat([positive_pairs, negative_pairs], ignore_index=True)

    # Keep unique article_id from item features
    item_df = articles_fg.read()
    item_df.drop_duplicates(subset="article_id", inplace=True)
    
    # Keep only the necessary columns from item features
    item_df = item_df[["article_id", "product_type_name", "product_group_name", "graphical_appearance_name", "colour_group_name", "perceived_colour_value_name", 
                       "perceived_colour_master_name", "department_name", "index_name", "index_group_name", "section_name", "garment_group_name"]]
    
    # Merge with item features
    ranking_df = ranking_df.merge(item_df, on="article_id")
    
    return ranking_df

ranking_df = compute_ranking_dataset(
    trans_fg, 
    articles_fg, 
    customers_fg,
)

ranking_df.head()

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Add Rankings Feature Group

rank_fg = fs.get_or_create_feature_group(
    name="ranking",
    version=2,
    description="Derived feature group for ranking",
    primary_key=["customer_id", "article_id"], 
    parents=[articles_fg, customers_fg, trans_fg],
)
rank_fg.insert(ranking_df)

feature_descriptions = [
    {"name": "customer_id", "description": "Unique identifier for each customer."},
    {"name": "article_id", "description": "Identifier for the purchased article."},
    {"name": "age", "description": "Age of the customer."},
    {"name": "month_sin", "description": "Sine of the month used for seasonal patterns."},
    {"name": "month_cos", "description": "Cosine of the month used for seasonal patterns."},
    {"name": "product_type_name", "description": "Name of the product type."},
    {"name": "product_group_name", "description": "Name of the product group."},
    {"name": "graphical_appearance_name", "description": "Name of the graphical appearance."},
    {"name": "colour_group_name", "description": "Name of the colour group."},
    {"name": "perceived_colour_value_name", "description": "Name of the perceived colour value."},
    {"name": "perceived_colour_master_name", "description": "Name of the perceived colour master."},
    {"name": "department_name", "description": "Name of the department."},
    {"name": "index_name", "description": "Name of the index."},
    {"name": "index_group_name", "description": "Name of the index group."},
    {"name": "section_name", "description": "Name of the section."},
    {"name": "garment_group_name", "description": "Name of the garment group."},
    {"name": "label", "description": "Label indicating whether the article was purchased (1) or not (0)."},
]

for desc in feature_descriptions: 
    rank_fg.update_feature_description(desc["name"], desc["description"])
