==> 02_retrieval_model_train.ipynb <==
<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Setup

<div style="background-color:teal; color:white; padding:5px; font-size:20px">
Installs

!pip install hopsworks

!pip install tensorflow

!pip install tensorboard

!pip install tensorflow_recommenders

!pip install tensorflow_addons

<div style="background-color:teal; color:white; padding:5px; font-size:20px">
Imports

import tensorflow as tf
from tensorflow import keras
from tensorflow.keras.layers.experimental.preprocessing import StringLookup, Normalization
import datetime
import warnings
warnings.filterwarnings('ignore')

# Tensorboard
import tensorboard
log_dir = "logs/train_logs/" + datetime.datetime.now().strftime("%Y%m%d-%H%M%S") #  Create Log Dir
summary_writer = tf.summary.create_file_writer(log_dir) # Instantiate writer

# Tensorboard notebook extension
%load_ext tensorboard

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Hopsworks Feature Store Connection

import boto3

# Initialize a client for Systems Manager
ssm = boto3.client('ssm', region_name='us-east-1')

parameter_name = 'hopsworks-api-key'

# Fetch the parameter
response = ssm.get_parameter(Name=parameter_name, WithDecryption=True)

# Extract the parameter value (API key in this case)
hopsworks_api_key = response['Parameter']['Value']

import hopsworks

project = hopsworks.login()

fs = project.get_feature_store()

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Training Dataset Creation

<div style="background-color:teal; color:white; padding:5px; font-size:20px">
Create Dataset From Feature Store (Hopsworks)

# Fetch Feature Groups
trans_fg = fs.get_feature_group(name="transactions",version=1)
customers_fg = fs.get_feature_group(name="customers",version=1)
articles_fg = fs.get_feature_group(name="articles",version=1)

# Join Desired Features to Define 'View'
selected_features = trans_fg.select(["customer_id", "article_id", "t_dat", "price", "month_sin", "month_cos"])\
    .join(customers_fg.select(["age", "club_member_status", "age_group"]), on="customer_id")\
    .join(articles_fg.select(["garment_group_name", "index_group_name"]), on="article_id")

selected_features.show(5)

# Create Retrieval Dataset View
feature_view = fs.get_or_create_feature_view(
    name='retrieval',
    query=selected_features,
    version=1,
)

# If View Already Created:
feature_view = fs.get_feature_view(name='retrieval', version=1)

# Fetch Datasets From View Using Hopsworks Train-test Split Function
train_df, val_df, test_df, _, _, _ = feature_view.train_validation_test_split(
    validation_size=0.1, 
    test_size=0.1,
    description='Retrieval dataset splits',
)

train_df["article_id"] = train_df["article_id"].astype(str)  # to be removed
val_df["article_id"] = val_df["article_id"].astype(str)  # to be removed

query_features = ["customer_id", "age", "month_sin", "month_cos"]
candidate_features = ["article_id", "garment_group_name", "index_group_name"]

def df_to_ds(df):
    # Convert df to tensoflow dataset
    return tf.data.Dataset.from_tensor_slices({col : df[col] for col in df})

BATCH_SIZE = 2048
train_ds = df_to_ds(train_df).batch(BATCH_SIZE).cache().shuffle(BATCH_SIZE*10)
val_ds = df_to_ds(val_df).batch(BATCH_SIZE).cache()

Creating unique-element lists for categorical embeddings

user_id_list = train_df["customer_id"].unique().tolist()
item_id_list = train_df["article_id"].unique().tolist()

garment_group_list = train_df["garment_group_name"].unique().tolist()
index_group_list = train_df["index_group_name"].unique().tolist()

print(f"Number of transactions: {len(train_df):,}")
print(f"Number of users: {len(user_id_list):,}")
print(f"Number of items: {len(item_id_list):,}")
print(garment_group_list)

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Retrieval Model

FactorizedTopK:
- Positive samples are only ones associated with the INDIVIDUAL query record. It does not reward for retrieving an item associated with the user on a separate record.

EMB_DIM = 16

<div style="background-color:teal; color:white; padding:5px; font-size:20px">
Query Tower Definition

class QueryTower(tf.keras.Model):

    def __init__(self):
        super().__init__()
        
        # user_embedding: Integer-Encoding
        self.user_embedding = tf.keras.Sequential([
            StringLookup(
                vocabulary=user_id_list,
                mask_token=None
            ),
            tf.keras.layers.Embedding(
                # We add an additional embedding to account for unknown tokens.
                len(user_id_list) + 1,
                EMB_DIM
            )
        ])

        # age: Normalization
        self.normalized_age = Normalization(axis=None)
        
        # Embedding Layer Definition
        self.fnn = tf.keras.Sequential([
            tf.keras.layers.Dense(EMB_DIM, activation="relu"),
            tf.keras.layers.Dense(EMB_DIM)
        ])

    def call(self, inputs):
        # Concat respectively transformed inputs
        concatenated_inputs = tf.concat([
            self.user_embedding(inputs["customer_id"]),
            tf.reshape(self.normalized_age(inputs["age"]), (-1,1)),
            tf.reshape(inputs["month_sin"], (-1,1)),
            tf.reshape(inputs["month_cos"], (-1,1))
        ], axis=1)

        # Send through embedding layer
        outputs = self.fnn(concatenated_inputs)

        return outputs

# Instantiate Model
query_model = QueryTower()

# Computing dataset-wide stats for normalization
query_model.normalized_age.adapt(train_ds.map(lambda x : x["age"]))

# Model Inputs:
query_df = train_df[query_features] # Subset to query feats
query_ds = df_to_ds(query_df).batch(1) # convert df to tf dataset

# load inputs into model
query_model(next(iter(query_ds))) 

<div style="background-color:teal; color:white; padding:5px; font-size:20px">
Item Tower Definition

class ItemTower(tf.keras.Model):

    def __init__(self):
        super().__init__()

        self.item_embedding = tf.keras.Sequential([
            StringLookup(
                vocabulary=item_id_list,
                mask_token=None
            ),
            tf.keras.layers.Embedding(
                # We add an additional embedding to account for unknown tokens.
                len(item_id_list) + 1,
                EMB_DIM
            )
        ])

        self.garment_group_tokenizer = StringLookup(vocabulary=garment_group_list, mask_token=None)
        self.index_group_tokenizer = StringLookup(vocabulary=index_group_list, mask_token=None)

        self.fnn = tf.keras.Sequential([
            tf.keras.layers.Dense(EMB_DIM, activation="relu"),
            tf.keras.layers.Dense(EMB_DIM)
        ])

    def call(self, inputs):
        garment_group_embedding = tf.one_hot(
            self.garment_group_tokenizer(inputs["garment_group_name"]),
            len(garment_group_list)
        )

        index_group_embedding = tf.one_hot(
            self.index_group_tokenizer(inputs["index_group_name"]),
            len(index_group_list)
        )

        concatenated_inputs = tf.concat([
            self.item_embedding(inputs["article_id"]),
            garment_group_embedding,
            index_group_embedding
        ], axis=1)

        outputs = self.fnn(concatenated_inputs)

        return outputs

# Instantiate Model
item_model = ItemTower()

# Model Inputs:
item_df = train_df[candidate_features] # Subset to item feats
item_df.drop_duplicates(subset="article_id", inplace=True) # Drop dup col
item_ds = df_to_ds(item_df) # Convert to tf dataset

<div style="background-color:teal; color:white; padding:5px; font-size:20px">
Two Tower Model Definition

import tensorflow_recommenders as tfrs

class TwoTowerModel(tf.keras.Model):
    def __init__(self, query_model, item_model):
        super().__init__()
        self.query_model = query_model
        self.item_model = item_model
        self.task = tfrs.tasks.Retrieval(
            
            metrics=tfrs.metrics.FactorizedTopK(
                # Grab the batch of items and retrieve embeddings
                candidates=item_ds.batch(BATCH_SIZE).map(self.item_model)
            )
        )

    def train_step(self, batch) -> tf.Tensor:
        # Set up a gradient tape to record gradients.
        with tf.GradientTape() as tape:

            # Loss computation.
            user_embeddings = self.query_model(batch)
            item_embeddings = self.item_model(batch)
            loss = self.task(
                user_embeddings, 
                item_embeddings,
                compute_metrics=False, # Skip summary statistics
            )

            # Handles regularization if it has been applied 
            regularization_loss = sum(self.losses)

            total_loss = loss + regularization_loss

        gradients = tape.gradient(total_loss, self.trainable_variables) # Calculate gradients
        self.optimizer.apply_gradients(zip(gradients, self.trainable_variables)) # update weights

        metrics = {
            "loss": loss,
            "regularization_loss": regularization_loss,
            "total_loss": total_loss
        }

        return metrics

    def test_step(self, batch) -> tf.Tensor:
        # Loss computation.
        user_embeddings = self.query_model(batch)
        item_embeddings = self.item_model(batch)

        loss = self.task(
            user_embeddings, 
            item_embeddings,
            compute_metrics=False,
        )

        # Handle regularization losses as well.
        regularization_loss = sum(self.losses)

        total_loss = loss + regularization_loss

        metrics = {metric.name: metric.result() for metric in self.metrics}
        metrics["loss"] = loss
        metrics["regularization_loss"] = regularization_loss
        metrics["total_loss"] = total_loss

        return metrics

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Train

import tensorflow_addons as tfa

model = TwoTowerModel(query_model, item_model)

optimizer = tfa.optimizers.AdamW(0.001, learning_rate=0.01)
model.compile(optimizer=optimizer)

# tf.keras.utils.plot_model(model)

# Define the Keras TensorBoard callback.
logdir="logs/fit/" + datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
tensorboard_callback = keras.callbacks.TensorBoard(log_dir=logdir)

# Train Model
model.fit(train_ds,
          validation_data=val_ds,
          epochs=5)

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Save Models

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Local Saves

The code below wraps the model in a module so that the raw input can be returned along with the embedding of a given input query.

class QueryModelModule(tf.Module):
    def __init__(self, query_model):
        self.query_model = query_model

    @tf.function() # Decorator allowing for input specification
    # Takes in user data and returns embedding + user data
    def compute_emb(self, instances):
        query_emb = self.query_model(instances)
        return {"customer_id": instances["customer_id"],
                "month_sin": instances["month_sin"],
                "month_cos": instances["month_cos"],
                "query_emb": query_emb}


# Wrap only the Query Tower model into new QueryModelModule
query_model = QueryModelModule(model.query_model)

# Define input specifications
instances_spec={
    'customer_id': tf.TensorSpec(shape=(None,), dtype=tf.string, name='customer_id'),
    'month_sin': tf.TensorSpec(shape=(None,), dtype=tf.float64, name='month_sin'),
    'month_cos': tf.TensorSpec(shape=(None,), dtype=tf.float64, name='month_cos'),
    'age': tf.TensorSpec(shape=(None,), dtype=tf.float64, name='age')
}
# Apply input specs to compute_embed()
signatures = query_model.compute_emb.get_concrete_function(instances_spec)

# Save ModelModule w/ signature
tf.saved_model.save(query_model, "model_artifacts/query_model", signatures=signatures)

# Save Model Locally
tf.saved_model.save(model.item_model, "model_artifacts/candidate_model")

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Save To Model Registry (Hopsworks)

import hsml

# Connect to Hopsworks Model Registry
mr = project.get_model_registry()

from hsml.schema import Schema
from hsml.model_schema import ModelSchema

# Infer input schema from data.
query_model_input_schema = Schema(query_df)

# Manually specify output schema.
query_model_output_schema = Schema([{
    "name": "query_embedding",
    "type": "float32",
    "shape": [EMB_DIM]
}])

query_model_schema = ModelSchema(
    input_schema=query_model_input_schema,
    output_schema=query_model_output_schema
)

query_model_schema.to_dict()

query_example = query_df.sample().to_dict("records")

mr_query_model = mr.tensorflow.create_model(
    name="query_model",
    description="Model that generates query embeddings from user and transaction features",
    input_example=query_example,
    model_schema=query_model_schema,
)
mr_query_model.save("query_model")

candidate_model_input_schema = Schema(item_df)

candidate_model_output_schema = Schema([{
    "name": "candidate_embedding",
    "type": "float32",
    "shape": [EMB_DIM],
}])

candidate_model_schema = ModelSchema(
    input_schema=candidate_model_input_schema,
    output_schema=candidate_model_output_schema
)

candidate_example = item_df.sample().to_dict("records")

mr_candidate_model = mr.tensorflow.create_model(
    name="candidate_model",
    description="Model that generates candidate embeddings from item features",
    input_example=candidate_example,
    model_schema=candidate_model_schema,
)
mr_candidate_model.save("candidate_model")
