==> 03_build-index.ipynb <==
<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Setup

!pip install hopsworks

!pip install tensorflow

!pip install pinecone-client

<div style="background-color:teal; color:white; padding:5px; font-size:20px">
Imports

import tensorflow as tf
import pandas as pd
import pprint
import numpy as np

import importlib
import sys
sys.path.append('./scripts')

import boto3

# Initialize a client for Systems Manager
ssm = boto3.client('ssm', region_name='us-east-1')

parameter_name = 'hopsworks-api-key'

# Fetch the parameter
response = ssm.get_parameter(Name=parameter_name, WithDecryption=True)

# Extract the parameter value (API key in this case)
hopsworks_api_key = response['Parameter']['Value']

# Pinecone Secrets
parameter_name = 'pinecone-recsys-api-key'
response = ssm.get_parameter(Name=parameter_name, WithDecryption=True)  # Fetch parameter
pinecone_recsys_api_key = response['Parameter']['Value']  # Extract parameter value

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Hopsworks Feature Store Connection

import hopsworks

project = hopsworks.login()

fs = project.get_feature_store()
mr = project.get_model_registry()

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Model & Data Imports

# Fetch Model From Registry
model = mr.get_model(
    name="candidate_model",
    version=1,
)

model_path = model.download()  # Set download path
candidate_model = tf.saved_model.load(model_path)  # Load Model

# Download Training Features 
feature_view = fs.get_feature_view(name="retrieval", version=1)

train_df, val_df, test_df, _, _, _ = feature_view.train_validation_test_split(
    validation_size=0.1, 
    test_size=0.1,
    description='Retrieval dataset splits',
)

train_df.head(3)

# Set cat. variables as strings
train_df["article_id"] = train_df["article_id"].astype(str)
val_df["article_id"] = val_df["article_id"].astype(str)

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Create Candidate Item Embeddings

# Get list of input features for the candidate model.
model_schema = model.model_schema['input_schema']['columnar_schema']
candidate_features = [feat['name'] for feat in model_schema]
candidate_features

# Get list of unique candidate items.
item_df = train_df[candidate_features]
item_df.drop_duplicates(subset="article_id", inplace=True)

# Convert to TF Dataset
item_ds = tf.data.Dataset.from_tensor_slices(
    {col: item_df[col] for col in item_df})

# Compute embeddings for all candidate items.
candidate_embeddings = item_ds.batch(2048).map(
    lambda x: (x["article_id"], candidate_model(x)))

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Prep Data For Upserting

# Concatenate all article IDs and embeddings from the candidate_embeddings dataset
all_article_ids = tf.concat([batch[0] for batch in candidate_embeddings], axis=0)
all_embeddings = tf.concat([batch[1] for batch in candidate_embeddings], axis=0)

# Convert tensors to numpy arrays
all_article_ids_np = all_article_ids.numpy().astype(int)
all_embeddings_np = all_embeddings.numpy()

# Convert numpy arrays to lists
items_ids_list = all_article_ids_np.tolist()
embeddings_list = all_embeddings_np.tolist()

# Create a list of dictionaries for upserting
to_upsert = [{"id": str(item_id), "values": embedding} for item_id, embedding in zip(items_ids_list, embeddings_list)]

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Upsert Into Pinecone Vector Database

from pinecone import Pinecone

pc = Pinecone(api_key=pinecone_recsys_api_key)
index = pc.Index("recsys-project")

pc.list_indexes()

import itertools
from pinecone import Pinecone


def chunks(iterable, batch_size=100):
    """A helper function to break an iterable into chunks of size batch_size."""
    it = iter(iterable)
    chunk = tuple(itertools.islice(it, batch_size))
    while chunk:
        yield chunk
        chunk = tuple(itertools.islice(it, batch_size))

# Upsert data with 100 vectors per upsert request
for i in chunks(to_upsert, batch_size=100):
    index.upsert(vectors=i) 

# check number of records in the index
index.describe_index_stats()

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Upload Embeddings to Hopsworks Feature Store

# Create a DataFrame
data_emb = pd.DataFrame({
    'article_id': items_ids_list, 
    'embeddings': embeddings_list,
})

data_emb.head()

# Create'candidate_embeddings_fg' feature group
candidate_embeddings_fg = fs.get_or_create_feature_group(
    name="candidate_embeddings_fg",
    primary_key=['article_id'],
    version=1,
    description='Embeddings for each article',
    online_enabled=True,
)

candidate_embeddings_fg.insert(data_emb)

# Create 'candidate_embeddings' feature view
feature_view = fs.get_or_create_feature_view(
    name="candidate_embeddings",
    version=1,
    description='Embeddings of each article',
    query=candidate_embeddings_fg.select(["article_id"]),
)
