==> 04_train-ranking-model.ipynb <==
<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Setup

!pip install hopsworks

!pip install catboost==1.1.1

<div style="background-color:teal; color:white; padding:5px; font-size:20px">
Imports

import pandas as pd
from catboost import CatBoostClassifier, Pool
from sklearn.metrics import classification_report, precision_recall_fscore_support
import joblib
from catboost import CatBoostClassifier, Pool

import importlib
import sys
sys.path.append('./scripts')

import boto3

# Initialize a client for Systems Manager
ssm = boto3.client('ssm', region_name='us-east-1')

parameter_name = 'hopsworks-api-key'

# Fetch the parameter
response = ssm.get_parameter(Name=parameter_name, WithDecryption=True)

# Extract the parameter value (API key in this case)
hopsworks_api_key = response['Parameter']['Value']

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Hopsworks Feature Store Connection

import hopsworks

project = hopsworks.login()

fs = project.get_feature_store()

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Data Import

# Fetch Feature Groups

customers_fg = fs.get_feature_group(name="customers", version=1)

articles_fg = fs.get_feature_group(name="articles", version=1)

trans_fg = fs.get_feature_group(name="transactions", version=1)

rank_fg = fs.get_feature_group(name="ranking", version=2)

'''Create Feature Views'''

# Customers
selected_features_customers = customers_fg.select_all()

fs.get_or_create_feature_view( 
    name='customers',
    query=selected_features_customers,
    version=1,
)

# Articles
selected_features_articles = articles_fg.select_all()

fs.get_or_create_feature_view(
    name='articles',
    query=selected_features_articles,
    version=1,
)

# Ranking
selected_features_ranking = rank_fg.select_except(["customer_id", "article_id"])

feature_view_ranking = fs.get_or_create_feature_view(
    name='ranking',
    query=selected_features_ranking,
    labels=["label"],
    version=2,
)

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Training Data

X_train, X_val, y_train, y_val = feature_view_ranking.train_test_split(
    test_size=0.1,
    description='Ranking training dataset',
)

X_train.head(3)

print(y_train['label'].unique())
y_train.head(3)

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Model Training

# Get Feature Cols
cat_features = list(
    X_train.select_dtypes(include=['string', 'object']).columns
)

# Convert to Catboost-optmized data structure
pool_train = Pool(X_train, y_train, cat_features=cat_features)
pool_val = Pool(X_val, y_val, cat_features=cat_features)

model = CatBoostClassifier(
    learning_rate=0.2,
    iterations=100,
    depth=10,
    scale_pos_weight=10,
    early_stopping_rounds=5,
    use_best_model=True,
)

model.fit(
    pool_train, 
    eval_set=pool_val,
)

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Model Validation

preds = model.predict(pool_val)

precision, recall, fscore, _ = precision_recall_fscore_support(y_val, preds, average="binary")

metrics = {
    "precision" : precision,
    "recall" : recall,
    "fscore" : fscore,
}
print(classification_report(y_val, preds))

FEATURE IMPORTANCE

feat_to_score = {
    feature: score 
    for feature, score 
    in zip(
        X_train.columns, 
        model.feature_importances_,
    )
}

feat_to_score = dict(
    sorted(
        feat_to_score.items(),
        key=lambda item: item[1],
        reverse=True,
    )
)
feat_to_score

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Save and Upload to Registry

joblib.dump(model, 'ranking_model.pkl')

# Connect to Hopsworks Model Registry
mr = project.get_model_registry()

from hsml.schema import Schema
from hsml.model_schema import ModelSchema

input_example = X_train.sample().to_dict("records")
input_schema = Schema(X_train)
output_schema = Schema(y_train)
model_schema = ModelSchema(input_schema, output_schema)

ranking_model = mr.python.create_model(
    name="ranking_model", 
    metrics=metrics,
    model_schema=model_schema,
    input_example=input_example,
    description="Ranking model that scores item candidates",
)
ranking_model.save("model_artifacts/ranking_model.pkl")
