==> 05_deployment.ipynb <==
<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Setup

!pip install hopsworks

<div style="background-color:teal; color:white; padding:0px; font-size:20px">
Imports

import os

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Hopsworks Feature Store Connection

import hopsworks

project = hopsworks.login()

fs = project.get_feature_store()

# Connect to Hopsworks Model Registry
mr = project.get_model_registry()

dataset_api = project.get_dataset_api()

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Ranking Transformer & Model Definition

ranking_model = mr.get_best_model(
    name="ranking_model", 
    metric="fscore", 
    direction="max",
)
ranking_model

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Transformer Definition

%%writefile scripts/ranking_transformer.py

# PINECONE SETUP
# - - - - - - - - - - - - - - - - - - 
import subprocess
import sys

# Use subprocess to run the pip install command
subprocess.check_call([sys.executable, "-m", "pip", "install", "pinecone-client"])

import pinecone
from pinecone import Pinecone

# IMPORTS
# - - - - - - - - - - - - - - - - - - 

import os
import pandas as pd

import hopsworks
from opensearchpy import OpenSearch

import logging

class Transformer(object):
    
    def __init__(self):
        
        # Connect to Hopsworks
        project = hopsworks.connection().get_project()
        self.fs = project.get_feature_store()

        # Connect to Pinecone
        self.pc = Pinecone(api_key='83447319-a1d1-446b-bfbb-9fbce3071957')
        
        # Establish the 'articles' feature view
        self.articles_fv = self.fs.get_feature_view(
            name="articles", 
            version=1,
        )
        
        # Get list of feature names for articles
        self.articles_features = [feat.name for feat in self.articles_fv.schema]
        
        # Establish the 'customers' feature view
        self.customer_fv = self.fs.get_feature_view(
            name="customers", 
            version=1,
        )

        # Establish the 'candidate_embeddings' feature view
        self.pc_index = self.pc.Index("recsys-project")

        # Retrieve ranking model
        mr = project.get_model_registry()
        model = mr.get_model(
            name="ranking_model", 
            version=1,
        )
        
        # Extract input schema from the model
        input_schema = model.model_schema["input_schema"]["columnar_schema"]
        
        # Get the names of features expected by the ranking model
        self.ranking_model_feature_names = [feat["name"] for feat in input_schema]

    def preprocess(self, inputs):
        # Extract the input instance
        inputs = inputs["instances"][0]
        
        # Extract customer_id from inputs
        customer_id = inputs["customer_id"]

        # Retrieve the 'candidate_embeddings' feature view
        neighbors = self.pc_index.query(
            vector=inputs["query_emb"],
            top_k=100,
            include_values=True)

        # Extract item_ids from returned object
        neighbors = [match['id'] for match in neighbors['matches']]

        # BUILD ARTICLE DATA INPUTS
        # - - - - - - - - - - - - - - - - - - - - - - - - - -  
        # Filter Out previously purchased candidates
        # Enrich remaining with Article data
        
        # Get IDs of items already bought by the customer
        already_bought_items_ids = self.fs.sql(
            f"SELECT article_id from transactions_1 WHERE customer_id = '{customer_id}'"
        ).values.reshape(-1).tolist()

        # Filter candidate items to exclude those already bought by the customer
        item_id_list = [
            str(item_id) 
            for item_id 
            in neighbors 
            if str(item_id) 
            not in already_bought_items_ids
        ]
        # Create df from filtered list
        item_id_df = pd.DataFrame({"article_id" : item_id_list})
        
        # Retrieve Article data for candidate items
        articles_data = [
            self.articles_fv.get_feature_vector({"article_id": item_id}) 
            for item_id 
            in item_id_list
        ]

        articles_df = pd.DataFrame(
            data=articles_data, 
            columns=self.articles_features,
        )
        
        # Join candidate items with their features
        ranking_model_inputs = item_id_df.merge(
            articles_df, 
            on="article_id", 
            how="inner",
        )        

        # ADD CUSTOMER DATA
        # - - - - - - - - - - - - - - - - - - - - - 
        
        # Add customer features to ranking_model_inputs
        customer_features = self.customer_fv.get_feature_vector(
            {"customer_id": customer_id}, 
            return_type="pandas",
        )
        ranking_model_inputs["age"] = customer_features.age.values[0]   
        ranking_model_inputs["month_sin"] = inputs["month_sin"]
        ranking_model_inputs["month_cos"] = inputs["month_cos"]

        # Select only the features required by the ranking model
        ranking_model_inputs = ranking_model_inputs[self.ranking_model_feature_names]
                
        return { 
            "inputs" : [{"ranking_features": ranking_model_inputs.values.tolist(), "article_ids": item_id_list}]
        }
        
    # A simple postprocess method that just returns the model's output
    def postprocess(self, outputs):
        return outputs

# Copy transformer file into Hopsworks File System 
uploaded_file_path = dataset_api.upload(
    "ranking_transformer.py",    # File name to be uploaded
    "Resources",                 # Destination directory in Hopsworks File System 
    overwrite=True,              # Overwrite the file if it already exists
) 

# Construct the path to the uploaded transformer script
transformer_script_path = os.path.join(
    "/Projects",                 # Root directory for projects in Hopsworks
    project.name,                # Name of the current project
    uploaded_file_path,          # Path to the uploaded file within the project
)
print(transformer_script_path)

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Ranking Predictor Definition

%%writefile scripts/ranking_predictor.py

import os
import joblib
import numpy as np

import logging

class Predict(object):
    
    def __init__(self):
        # Instantiate preloaded model through env. path
        self.model = joblib.load(os.environ["ARTIFACT_FILES_PATH"] + "/ranking_model.pkl")

    def predict(self, inputs):
        # Extract ranking features and article IDs from the inputs
        features = inputs[0].pop("ranking_features")
        article_ids = inputs[0].pop("article_ids")
        
        # Log the extracted features
        logging.info("predict -> " + str(features))

        # Predict probabilities for the positive class
        scores = self.model.predict_proba(features).tolist()
        
        # Get scores of positive class
        scores = np.asarray(scores)[:,1].tolist() 

        # Return the predicted scores along with the corresponding article IDs
        return {
            "scores": scores, 
            "article_ids": article_ids,
        }

# Upload predictor file to Hopsworks
uploaded_file_path = dataset_api.upload(
    "ranking_predictor.py", 
    "Resources", 
    overwrite=True,
)

# Construct the path to the uploaded script
predictor_script_path = os.path.join(
    "/Projects", 
    project.name, 
    uploaded_file_path,
)

<div style="background-color:teal; color:white; padding:0px; font-size:15px">
Deploy Ranking Transformer & Predictor

from hsml.transformer import Transformer

ranking_deployment_name = "rankingdeployment"

# Define transformer
ranking_transformer=Transformer(
    script_file=transformer_script_path, 
    resources={"num_instances": 0},
)

# Deploy ranking model
ranking_deployment = ranking_model.deploy(
    name=ranking_deployment_name,
    description="Deployment that search for item candidates and scores them based on customer metadata",
    script_file=predictor_script_path,
    resources={"num_instances": 0},
    transformer=ranking_transformer,
)

# Start the deployment
ranking_deployment.start()

# Check logs in case of failure
# ranking_deployment.get_logs(component="predictor", tail=200)

<div style="background-color:teal; color:white; padding:0px; font-size:15px">
Test Input

# Define a test input example
test_ranking_input = {"instances": [{"customer_id": "641e6f3ef3a2d537140aaa0a06055ae328a0dddf2c2c0dd6e60eb0563c7cbba0",
    "month_sin": 1.2246467991473532e-16,
    "query_emb": [0.214135289,
     0.571055949,
     0.330709577,
     -0.225899458,
     -0.308674961,
     -0.0115124583,
     0.0730511621,
     -0.495835781,
     0.625569344,
     -0.0438038409,
     0.263472944,
     -0.58485353,
     -0.307070434,
     0.0414443575,
     -0.321789205,
     0.966559],
    "month_cos": -1.0}]}

# Test ranking deployment
ranked_candidates = ranking_deployment.predict(test_ranking_input)

print(len(ranked_candidates['predictions']['scores']))
ranked_candidates['predictions']['scores']

# Check logs in case of failure
# ranking_deployment.get_logs(component="transformer",tail=200)

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Query Model

# Retrieve the 'query_model' from the Model Registry
query_model = mr.get_model(
    name="query_model",
    version=1,
)

<div style="background-color: darkgreen ; color:white; padding:0px; font-size:15px">
Query Transformer Definition

%%writefile scripts/querymodel_transformer.py

import os
import numpy as np
import pandas as pd
from datetime import datetime

import hopsworks

import logging

class Transformer(object):
    
    def __init__(self):            
        # Connect to the Hopsworks
        project = hopsworks.connection().get_project()
        ms = project.get_model_serving()
    
        # Retrieve the 'customers' feature view
        fs = project.get_feature_store()
        self.customer_fv = fs.get_feature_view(
            name="customers", 
            version=1,
        )
        # Retrieve the ranking deployment 
        self.ranking_server = ms.get_deployment("rankingdeployment")

    def preprocess(self, inputs):
        # Check if the input data contains a key named "instances"
        # and extract the actual data if present
        inputs = inputs["instances"] if "instances" in inputs else inputs
        
        # Extract customer_id and transaction_date from the inputs
        customer_id = inputs["customer_id"]
        transaction_date = inputs["transaction_date"]
        
        # Extract month from the transaction_date
        month_of_purchase = datetime.fromisoformat(inputs.pop("transaction_date"))
        
        # Get customer features
        customer_features = self.customer_fv.get_feature_vector(
            {"customer_id": customer_id}, 
            return_type="pandas",
        )
        
        # Enrich inputs with customer age
        inputs["age"] = customer_features.age.values[0]   
        
        # Calculate the sine and cosine of the month_of_purchase
        month_of_purchase = datetime.strptime(transaction_date, "%Y-%m-%dT%H:%M:%S.%f").month
        
        # Calculate a coefficient for adjusting the periodicity of the month
        coef = np.random.uniform(0, 2 * np.pi) / 12
        
        # Calculate the sine and cosine components for the month_of_purchase
        inputs["month_sin"] = float(np.sin(month_of_purchase * coef)) 
        inputs["month_cos"] = float(np.cos(month_of_purchase * coef))
                
        return {
            "instances" : [inputs]
        }

    def postprocess(self, outputs):
        # Return ordered ranking predictions        
        return {
            "predictions": self.ranking_server.predict({ "instances": outputs["predictions"]}),
        }

# Copy transformer file into Hopsworks File System
uploaded_file_path = dataset_api.upload(
    "querymodel_transformer.py", 
    "Models", 
    overwrite=True,
)

# Construct the path to the uploaded script
transformer_script_path = os.path.join(
    "/Projects", 
    project.name, 
    uploaded_file_path,
)

from hsml.transformer import Transformer

query_model_deployment_name = "querydeployment"

# Define transformer
query_model_transformer=Transformer(
    script_file=transformer_script_path, 
    resources={"num_instances": 0},
)

# Deploy the query model
query_model_deployment = query_model.deploy(
    name=query_model_deployment_name,
    description="Deployment that generates query embeddings from customer and item features using the query model",
    resources={"num_instances": 0},
    transformer=query_model_transformer,
)

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Run Registered Deployment

# Start the deployment
query_model_deployment.start()

query_model_deployment.get_state()

# Check logs in case of failure
# query_model_deployment.get_logs(component="transformer", tail=20)

<div style="background-color:teal; color:white; padding:0px; font-size:15px">
Test Input

# Define a test input example
data = {"instances": {"customer_id": "641e6f3ef3a2d537140aaa0a06055ae328a0dddf2c2c0dd6e60eb0563c7cbba0", "transaction_date": "2022-11-15T12:16:25.330916"}}

# Test the deployment
ranked_candidates = query_model_deployment.predict(data)

# Check logs in case of failure
# query_model_deployment.get_logs(component="transformer",tail=200)

ranked_candidates

<div style="background-color:moccasin; color:black; padding:5px; font-size:20px">
Stop Deployment

# Stop the ranking model deployment
ranking_deployment.stop()

# Stop the query model deployment
query_model_deployment.stop()
